utils: move maybe_int to helpers
